+++ src/app.module.ts
@@ -1,12 +1,11 @@
 import { Module } from "@nestjs/common";
 
-import { OrdersModule } from "@orders/orders.module";
-
-import { ProductsModule } from "./products/products.module";
-import { ReviewsModule } from "./reviews/reviews.module";
-import { ShoppingBasketModule } from "./shopping-basket/shopping-basket.module";
-import { StoreModule } from "./store/store.module";
-import { UserModule } from "./user/user.module";
+import { OrdersModule } from "@root/modules/orders/orders.module";
+import { ProductsModule } from "@root/modules/products/products.module";
+import { ReviewsModule } from "@root/modules/reviews/reviews.module";
+import { ShoppingBasketModule } from "@root/modules/shopping-basket/shopping-basket.module";
+import { StoreModule } from "@root/modules/store/store.module";
+import { UserModule } from "@root/modules/user/user.module";
 
 @Module({
   imports: [UserModule, StoreModule, ProductsModule, OrdersModule, ReviewsModule, ShoppingBasketModule],
