+++ src/modules/orders/services/confirm-order/confirm-order.interface.ts
@@ -1,4 +1,4 @@
-import { ConfirmOrdersDto } from "@orders/dto/request/confirm-orders.dto";
+import { ConfirmOrdersDto } from "@root/modules/orders/dto/request/confirm-orders.dto";
 
 export interface ConfirmOrder {
   execute: (confirmOrdersDto: ConfirmOrdersDto) => Promise<string>;
