+++ src/modules/orders/services/confirm-order/confirm-order.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@nestjs/common";
 
-import { ConfirmOrdersDto } from "@orders/dto/request/confirm-orders.dto";
-import { ConfirmOrder } from "@orders/services/confirm-order/confirm-order.interface";
+import { ConfirmOrdersDto } from "@root/modules/orders/dto/request/confirm-orders.dto";
+import { ConfirmOrder } from "@root/modules/orders/services/confirm-order/confirm-order.interface";
 
 @Injectable()
 export class ConfirmOrderService implements ConfirmOrder {
