+++ src/modules/products/products.entity.ts
