+++ src/modules/store/store.entity.ts
