+++ src/modules/user/dto/update-user.dto.ts
@@ -1,4 +1,3 @@
-import { Transform } from "class-transformer";
 import { IsNotEmpty } from "class-validator";
 
 export class UpdateUserDto {
