+++ src/modules/user/service/user.interface.ts
