+++ test/products/jest-e2e.json
@@ -7,7 +7,7 @@
     "^.+\\.(t|j)s$": "ts-jest"
   },
   "moduleNameMapper": {
-    "^@orders/(.*)$": "<rootDir>/../../src/orders/$1",
+    "^@orders/(.*)$": "<rootDir>/../../src/modules/orders/$1",
     "^@root/(.*)$": "<rootDir>/../../src/$1"
   }
 }
