+++ test/user/user.e2e-spec.ts
@@ -4,7 +4,6 @@ import { Test, TestingModule } from "@nestjs/testing";
 import * as request from "supertest";
 
 import { AppModule } from "../../src/app.module";
-import { log } from "console";
 
 describe("user (e2e)", () => {
   let app: INestApplication;
